+++ src/main/java/com/strangegrotto/wealthdraft/Main.java
@@ -9,7 +9,6 @@ import ch.qos.logback.core.ConsoleAppender;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.type.CollectionType;
 import com.fasterxml.jackson.databind.type.MapType;
 import com.fasterxml.jackson.databind.type.TypeFactory;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -18,8 +17,9 @@ import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Ordering;
-import com.strangegrotto.wealthdraft.assetallocation.AssetAllocationRenderer;
-import com.strangegrotto.wealthdraft.assetallocation.TargetAssetAllocation;
+import com.strangegrotto.wealthdraft.assetallocation.calculator.AssetAllocationCalculator;
+import com.strangegrotto.wealthdraft.assetallocation.renderer.AssetAllocationRenderer;
+import com.strangegrotto.wealthdraft.assetallocation.datamodel.TargetAssetAllocations;
 import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitions;
 import com.strangegrotto.wealthdraft.errors.ValOrGerr;
 import com.strangegrotto.wealthdraft.govconstants.GovConstantsForYear;
@@ -49,6 +49,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.time.Year;
 import java.util.*;
@@ -81,8 +82,9 @@ public class Main {
     // TODO Make this months instead
     private static final int PROJECTION_DISPLAY_INCREMENT_YEARS = 1;
 
-    private static final double ASSET_ALLOCATION_DEVIATION_PCT_WARN = 0.10;
-    private static final double ASSET_ALLOCATION_DEVIATION_PCT_ERROR = 0.15;
+    // TODO Make asset allocation deviation thresholds configurable
+    private static final BigDecimal ASSET_ALLOCATION_DEVIATION_PCT_WARN = BigDecimal.valueOf(0.10);
+    private static final BigDecimal ASSET_ALLOCATION_DEVIATION_PCT_ERROR = BigDecimal.valueOf(0.15);
 
     private static final Logger log = LoggerFactory.getLogger(Main.class);
 
@@ -188,10 +190,9 @@ public class Main {
 
         String assetAllocationsFilepath = parsedArgs.getString(ASSET_ALLOCATIONS_FILEPATH_ARG);
         log.debug("Asset allocations filepath: {}", assetAllocationsFilepath);
-        CollectionType targetAssetAllocationsType = typeFactory.constructCollectionType(ArrayList.class, TargetAssetAllocation.class);
-        List<TargetAssetAllocation> targetAssetAllocations;
+        TargetAssetAllocations targetAssetAllocations;
         try {
-            targetAssetAllocations = mapper.readValue(new File(assetAllocationsFilepath), targetAssetAllocationsType);
+            targetAssetAllocations = mapper.readValue(new File(assetAllocationsFilepath), TargetAssetAllocations.class);
         } catch (IOException e) {
             log.error("An error occurred parsing the asset allocations file '{}'", assetAllocationsFilepath, e);
             System.exit(FAILURE_EXIT_CODE);
@@ -243,19 +244,20 @@ public class Main {
             System.exit(FAILURE_EXIT_CODE);
         }
 
-        var assetAllocationRenderer = new AssetAllocationRenderer(
-                display,
-                ASSET_ALLOCATION_DEVIATION_PCT_WARN,
-                ASSET_ALLOCATION_DEVIATION_PCT_ERROR
-        );
         var assetsHistoryByDate = assetsHistory.getHistoryByDate();
         var latestDate = assetsHistoryByDate.lastKey();
         var latestAssetSnapshots = assetsHistoryByDate.get(latestDate);
-        assetAllocationRenderer.renderAssetAllocations(
+        var assetAllocationCalculator = new AssetAllocationCalculator(
+                ASSET_ALLOCATION_DEVIATION_PCT_WARN,
+                ASSET_ALLOCATION_DEVIATION_PCT_ERROR
+        );
+        var assetAllocationCalcResults = assetAllocationCalculator.calculate(
                 targetAssetAllocations,
                 assetDefinitions.getAssets(),
                 latestAssetSnapshots
         );
+        var assetAllocationRenderer = new AssetAllocationRenderer(display);
+        assetAllocationRenderer.render(assetAllocationCalcResults);
     }
 
     @VisibleForTesting